refactor

+++ src/main/kotlin/com/almasb/zeph/character/components/CharacterComponent.kt
@@ -143,7 +143,7 @@ open class CharacterComponent(data: CharacterData) : Component() {
 
     private fun agi()   = attributes.getTotalAttribute(AGILITY)
 
-    private fun int_()  = attributes.getTotalAttribute(INTELLECT)
+    private fun int()  = attributes.getTotalAttribute(INTELLECT)
 
     private fun wis()   = attributes.getTotalAttribute(WISDOM)
 
@@ -176,7 +176,7 @@ open class CharacterComponent(data: CharacterData) : Component() {
         }, vit, str, level))
 
         stats.statProperty(MAX_SP).bind(createDoubleBinding(Callable {
-            1.0 + wis() * 0.4 + wil() * 0.3 + level() * 0.25 + (wis() / 10).toDouble() + int_() * 0.3 + charClass.value.sp * level()
+            1.0 + wis() * 0.4 + wil() * 0.3 + level() * 0.25 + (wis() / 10).toDouble() + int() * 0.3 + charClass.value.sp * level()
         }, wis, wil, level, int))
 
         stats.statProperty(HP_REGEN).bind(createDoubleBinding(Callable{ 1 + vit() * 0.1 },
@@ -188,19 +188,19 @@ open class CharacterComponent(data: CharacterData) : Component() {
         stats.statProperty(ATK).bind(createDoubleBinding(Callable{ str() * 0.5 + dex() * 0.3 + per() * 0.2 + luc() * 0.1 + level().toDouble() + (str() / 10 * (str() / 10 + 1)).toDouble() },
                 str, dex, per, luc, level))
 
-        stats.statProperty(MATK).bind(createDoubleBinding(Callable{ int_() * 0.5 + wis() * 0.4 + wil() * 0.4 + dex() * 0.3 + per() * 0.2 + luc() * 0.1 },
+        stats.statProperty(MATK).bind(createDoubleBinding(Callable{ int() * 0.5 + wis() * 0.4 + wil() * 0.4 + dex() * 0.3 + per() * 0.2 + luc() * 0.1 },
                 int, dex, per, luc))
 
         stats.statProperty(DEF).bind(createDoubleBinding(Callable{ vit() * 0.5 + per() * 0.2 + str() * 0.1 + level() * 0.25 + (vit() / 20).toDouble() },
                 vit, per, str, level))
 
-        stats.statProperty(MDEF).bind(createDoubleBinding(Callable{ wil() * 0.5 + wis() * 0.3 + per() * 0.2 + int_() * 0.1 + level() * 0.25 + (wil() / 20 * int_() / 10).toDouble() },
+        stats.statProperty(MDEF).bind(createDoubleBinding(Callable{ wil() * 0.5 + wis() * 0.3 + per() * 0.2 + int() * 0.1 + level() * 0.25 + (wil() / 20 * int() / 10).toDouble() },
                 wil, wis, per, int, level))
 
         stats.statProperty(ASPD).bind(createDoubleBinding(Callable{ agi() * 0.5 + dex() * 0.2 },
                 agi, dex))
 
-        stats.statProperty(MSPD).bind(createDoubleBinding(Callable{ dex() * 0.3 + wil() * 0.1 + wis() * 0.1 + int_() * 0.1 + per() * 0.1 + luc() * 0.1 },
+        stats.statProperty(MSPD).bind(createDoubleBinding(Callable{ dex() * 0.3 + wil() * 0.1 + wis() * 0.1 + int() * 0.1 + per() * 0.1 + luc() * 0.1 },
                 dex, wil, wis, int, per, luc))
 
         stats.statProperty(CRIT_CHANCE).bind(createDoubleBinding(Callable{ luc() * 0.5 + per() * 0.1 + wis() * 0.1 },
